+++ __tests__/__integration__/imperative/__tests__/__integration__/cli/profiles/cli.imperative-test-cli.profiles.create.secured-profile.integration.test.ts
@@ -31,7 +31,8 @@ describe("imperative-test-cli profiles create secured-profile", () => {
         const profileName: string = "my_secret";
         const response = runCliScript(__dirname + "/__scripts__/secured-profile/create_and_list.sh", TEST_ENVIRONMENT.workingDir,
             [secret, profileName]);
-        expect(response.stderr.toString()).toBe("");
+        expect(response.stderr.toString()).toContain("command 'profiles create' is deprecated");
+        expect(response.stderr.toString()).toContain("command 'profiles list' is deprecated");
         expect(response.status).toBe(0);
         expect(response.stdout.toString()).not.toContain(secret);
         expect(response.stdout.toString()).toContain(profileName);