+++ packages/cmd/src/CommandProcessor.ts
@@ -35,6 +35,7 @@ import { ImperativeConfig, TextUtils } from "../../utilities";
 import * as nodePath from "path";
 import * as os from "os";
 import { ICommandHandlerRequire } from "./doc/handler/ICommandHandlerRequire";
+import { IHandlerParameters } from "../../cmd";
 import { ChainedHandlerService } from "./ChainedHandlerUtils";
 import { Constants } from "../../constants";
 import { ICommandArguments } from "./doc/args/ICommandArguments";
@@ -43,7 +44,6 @@ import { WebHelpManager } from "./help/WebHelpManager";
 import { ICommandOptionDefinition } from "./doc/option/ICommandOptionDefinition";
 import { ICommandProfile } from "./doc/profiles/definition/ICommandProfile";
 import { Config } from "../../config";
-
 /**
  * The command processor for imperative - accepts the command definition for the command being issued (and a pre-built)
  * response object and validates syntax, loads profiles, instantiates handlers, & invokes the handlers.
@@ -563,18 +563,21 @@ export class CommandProcessor {
                 return this.finishResponse(response);
             }
 
+            const handlerParms: IHandlerParameters = {
+                response,
+                profiles: prepared.profiles,
+                arguments: prepared.args,
+                positionals: prepared.args._,
+                definition: this.definition,
+                fullDefinition: this.fullDefinition
+            };
             try {
-                await handler.process({
-                    response,
-                    profiles: prepared.profiles,
-                    arguments: prepared.args,
-                    positionals: prepared.args._,
-                    definition: this.definition,
-                    fullDefinition: this.fullDefinition
-                });
+                await handler.process(handlerParms);
             } catch (processErr) {
 
                 this.handleHandlerError(processErr, response, this.definition.handler);
+
+                CliUtils.showMsgWhenDeprecated(handlerParms);
 
                 // Return the failed response to the caller
                 return this.finishResponse(response);
@@ -589,6 +592,8 @@ export class CommandProcessor {
                 timingApi.mark("END_CMD_INVOKE");
                 timingApi.measure("Command executed: " + this.commandLine, "START_CMD_INVOKE", "END_CMD_INVOKE");
             }
+
+            CliUtils.showMsgWhenDeprecated(handlerParms);
 
             // Return the response to the caller
             return this.finishResponse(response);